+++ templates/common/ios/{%- project.name %}/AppDelegate.m
@@ -51,7 +51,7 @@ static void InitializeFlipper(UIApplication *application) {
   if (@available(iOS 13.0, *)) {
       rootView.backgroundColor = [UIColor systemBackgroundColor];
   } else {
-      rootView.backgroundColor = [UIColor blackColor];
+      rootView.backgroundColor = [UIColor whiteColor];
   }
 
   self.window = [[UIWindow alloc] initWithFrame:[UIScreen mainScreen].bounds];
