+++ templates/common/scripts/postinstall.sh
@@ -9,3 +9,5 @@ if [[ "$OSTYPE" == "darwin"* ]]; then
   echo "Installing pods..."
   cd ios && pod install --repo-update && cd ..;
 fi
+
+git add . && git commit -m 'Initial commit';
